requirements.txt
numpy
streamlit
streamlit-option-menu
pandas
yfinance
ephem
pystan
fbprophet
Pillow
